% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cate_blp_bootstrap.R
\name{cate_blp_bootstrap}
\alias{cate_blp_bootstrap}
\title{Bootstrapped predictions of the CATE-BLP}
\usage{
cate_blp_bootstrap(
  obj,
  iterations = 100,
  output = "totals",
  tau.treatment.baseline = 0
)
}
\arguments{
\item{obj}{\code{cateblp} class object created by \code{cate_blp}}

\item{iterations}{integer number of bootstrap iterations}

\item{output}{character of either \code{"totals"} or \code{"raw.list"}. If \code{"totals"}
then the total predicted treatment effect for each boot strap iteration is found
using both complete treatment (all units treated) or treatment only when the
predicted effect is greater than \code{tau.treatment.baseline}.}

\item{tau.treatment.baseline}{numeric. Treatment is only made for units when the
predicted treatment effect is > \code{tau.treatment.baseline}.}
}
\description{
Bootstrapped predictions of the CATE-BLP
}
