% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{predict}
\alias{predict}
\title{Predict method for cateblp objects}
\usage{
predict(obj, ...)
}
\arguments{
\item{obj}{cateblp object from \code{cate_blp}}

\item{...}{other}
}
\description{
Predict method for cateblp objects
}
