% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{summary}
\alias{summary}
\title{Summary method for cateblp objects}
\usage{
summary(obj, ...)
}
\arguments{
\item{obj}{cateblp object from \code{cate_blp}}

\item{...}{other}
}
\description{
Summary method for cateblp objects
}
