% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cate_blp.r
\name{cate_blp}
\alias{cate_blp}
\title{Estimates the best linear projection of the conditional average treatment
effect (CATE-BLP) with respect to a feature
vector A using cross-fitted DR/AIPW-scores.}
\usage{
cate_blp(fit, A)
}
\arguments{
\item{fit}{fitted model object output from \code{causal_forest}
in the \strong{grf} package or \code{dr_learner} from the \strong{drlearner} package.}

\item{A}{numeric matrix of features with the same number of rows as
the number of predictions from \code{fit}.}
}
\description{
Estimates the best linear projection of the conditional average treatment
effect (CATE-BLP) with respect to a feature
vector A using cross-fitted DR/AIPW-scores.
}
\references{
Semenova, V. and Chernozhukov, V., 2021. Debiased machine learning of
conditional average treatment effects and other causal functions.
The Econometrics Journal, 24(2), pp.264-289.
}
