% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cate_blp.r
\name{cate_blp}
\alias{cate_blp}
\title{Estimates the best linear projection of the conditional average treatment
effect (CATE-BLP) with respect to a feature
vector A using cross-fitted DR/AIPW-scores.}
\usage{
cate_blp(cf, A)
}
\arguments{
\item{cf}{\emph{causal_forest} object output from the \code{causal_forest} function
from the \strong{grf} package.}

\item{A}{numeric vector of features with the same number of rows as
\code{cf.predictions}}
}
\description{
Estimates the best linear projection of the conditional average treatment
effect (CATE-BLP) with respect to a feature
vector A using cross-fitted DR/AIPW-scores.
}
